chore: upgrade TypeScript to 6.0.2 and typescript-eslint to 8.58.2
Bumps TypeScript from 5.x to 6.0.2 across all workspaces (lib, server,
client, update-sidecar, agent-sidecar) and typescript-eslint to 8.58.2
in client and server. TS 6 is stricter, so this also includes the
minimal compatibility fixes needed to build cleanly:

- Add ignoreDeprecations: "6.0" to tsconfigs still using
  moduleResolution: "node" or baseUrl.
- Set explicit rootDir on server/tsconfig.json (TS 6 no longer infers
  it when includes span parent directories).
- Move the express-serve-static-core module augmentation out of lib/
  (which has no express dependency) into server/src/types/express.d.ts
  where the module is actually resolvable.
- Replace NodeJS.Timeout with ReturnType<typeof setTimeout> in client
  hooks so browser code doesn't depend on @types/node leaking through.
- Clean up a few pre-existing lint errors surfaced by the new toolchain
  (unused HealthStatus imports, {} type in system-settings route).
