+++ agent-sidecar/package.json
@@ -23,7 +23,7 @@
     "@types/express": "^5.0.3",
     "@types/node": "^24.12.2",
     "tsx": "^4.21.0",
-    "typescript": "^5.1.6",
+    "typescript": "^6.0.2",
     "vitest": "^4.1.3"
   }
 }
