+++ client/package.json
@@ -63,7 +63,6 @@
   },
   "devDependencies": {
     "@eslint/js": "^10.0.1",
-    "js-yaml": "^4.1.1",
     "@types/node": "^25.6.0",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
@@ -72,9 +71,10 @@
     "eslint-plugin-react-hooks": "^7.0.1",
     "eslint-plugin-react-refresh": "^0.5.2",
     "globals": "^17.5.0",
+    "js-yaml": "^4.1.1",
     "tw-animate-css": "^1.4.0",
-    "typescript": "~5.8.3",
-    "typescript-eslint": "^8.58.1",
+    "typescript": "^6.0.2",
+    "typescript-eslint": "^8.58.2",
     "vite": "^8.0.8"
   }
 }
