+++ lib/package.json
@@ -12,6 +12,6 @@
   },
   "devDependencies": {
     "@types/express": "^5.0.3",
-    "typescript": "^5.1.6"
+    "typescript": "^6.0.2"
   }
 }
