+++ server/package.json
@@ -39,8 +39,8 @@
         "@types/passport-google-oauth20": "^2.0.17",
         "@types/pg": "^8.20.0",
         "@types/supertest": "^7.2.0",
-        "@typescript-eslint/eslint-plugin": "^8.58.1",
-        "@typescript-eslint/parser": "^8.58.1",
+        "@typescript-eslint/eslint-plugin": "^8.58.2",
+        "@typescript-eslint/parser": "^8.58.2",
         "@vitest/coverage-v8": "^4.1.4",
         "cross-env": "^10.1.0",
         "eslint": "^10.2.0",
@@ -50,8 +50,8 @@
         "rimraf": "^6.1.3",
         "supertest": "^7.2.2",
         "tsx": "^4.21.0",
-        "typescript": "^5.9.3",
-        "typescript-eslint": "^8.58.1",
+        "typescript": "^6.0.2",
+        "typescript-eslint": "^8.58.2",
         "vitest": "^4.1.4"
     },
     "dependencies": {
