+++ server/src/services/haproxy/haproxy-service.ts
@@ -9,7 +9,6 @@ import {
   ServiceMetadata,
   ServiceStatusInfo,
   StartupResult,
-  HealthStatus,
   HealthStatusValues,
   NetworkRequirement,
   VolumeRequirement,
