+++ server/src/services/monitoring/monitoring-service.ts
@@ -11,7 +11,6 @@ import {
   ServiceMetadata,
   ServiceStatusInfo,
   StartupResult,
-  HealthStatus,
   HealthStatusValues,
   NetworkRequirement,
   VolumeRequirement,
