+++ server/tsconfig.json
@@ -3,6 +3,7 @@
     "target": "ES2022",
     "module": "CommonJS",
     "moduleResolution": "node",
+    "ignoreDeprecations": "6.0",
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
     "allowImportingTsExtensions": false,
@@ -10,6 +11,7 @@
     "isolatedModules": true,
     "noEmit": false,
     "outDir": "./dist",
+    "rootDir": "..",
     "strict": true,
     "types": ["vitest/globals"],
     "skipLibCheck": true,
