+++ update-sidecar/package.json
@@ -18,6 +18,6 @@
     "@types/dockerode": "^4.0.1",
     "@types/node": "^24.12.2",
     "tsx": "^4.21.0",
-    "typescript": "^5.1.6"
+    "typescript": "^6.0.2"
   }
 }
